styles: prettier end line at 160

+++ src/modules/users/services/users/create-user/create-user.service.ts
@@ -1,10 +1,7 @@
 import { ProfessionalRepository } from './../../../repositories/implementations/ProfessionalRepository';
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { InviteUsersRepository } from '../../../../../modules/users/repositories/implementations/InviteUsersRepository';
-import {
-  ErrorAuthEnum,
-  ErrorInvitesEnum,
-} from '../../../../../shared/constants/enum/errorMessage';
+import { ErrorAuthEnum, ErrorInvitesEnum } from '../../../../../shared/constants/enum/errorMessage';
 
 import { DayJSProvider } from '../../../../../shared/providers/DateProvider/implementations/DayJSProvider';
 import { HashProvider } from '../../../../../shared/providers/HashProvider/implementations/HashProvider';
@@ -26,18 +23,11 @@ export class CreateUserService {
 
   async execute({ token, password, ...restCreateUserDto }: CreateUserDto) {
     const passHash = await this.hashProvider.createHash(password);
-    const userAlreadyExist = await this.userRepository.findByEmail(
-      restCreateUserDto.email,
-    );
+    const userAlreadyExist = await this.userRepository.findByEmail(restCreateUserDto.email);
 
-    if (userAlreadyExist)
-      throw new BadRequestException(ErrorAuthEnum.USER_ALREADY_EXIST);
+    if (userAlreadyExist) throw new BadRequestException(ErrorAuthEnum.USER_ALREADY_EXIST);
 
-    const { companies, companyId, invite } = await getCompanyPermissionByToken(
-      token,
-      this.findByTokenService,
-      this.dateProvider,
-    );
+    const { companies, companyId, invite } = await getCompanyPermissionByToken(token, this.findByTokenService, this.dateProvider);
 
     const userData = {
       ...restCreateUserDto,
@@ -46,28 +36,16 @@ export class CreateUserService {
 
     const professional = invite?.professional;
 
-    const user = await this.userRepository.create(
-      userData,
-      companies,
-      professional,
-    );
+    const user = await this.userRepository.create(userData, companies, professional);
 
-    if (invite && invite.email && companyId)
-      await this.inviteUsersRepository.deleteByCompanyIdAndEmail(
-        companyId,
-        invite.email,
-      );
+    if (invite && invite.email && companyId) await this.inviteUsersRepository.deleteByCompanyIdAndEmail(companyId, invite.email);
 
     delete user.password;
     return user;
   }
 }
 
-export const getCompanyPermissionByToken = async (
-  token: string,
-  findByTokenService: FindByTokenService,
-  dateProvider: DayJSProvider,
-) => {
+export const getCompanyPermissionByToken = async (token: string, findByTokenService: FindByTokenService, dateProvider: DayJSProvider) => {
   if (!token) return { companies: [] };
 
   const invite = await findByTokenService.execute(token);
